UI/requirements.txt
Flask==3.0.3
pillow==11.0.0
PyMuPDF==1.24.13
PyPDF2==3.0.1